Moved Settings component into pages where it belongs
