+++ client/src/pages/Settings.js
@@ -1,6 +1,4 @@
 import React, { Component } from 'react'
-import logo from './../../logo.svg'
-import './../../App.css'
 
 class Settings extends Component {
   constructor (props) {
@@ -40,13 +38,6 @@ class Settings extends Component {
   render () {
     return (
       <>
-        <div className='App'>
-          <div className='App-header'>
-            <img src={logo} className='App-logo' alt='logo' />
-            <h2>Mildly Sharp Coders at Work!</h2>
-          </div>
-        </div>
-
         <div>
           <h1>Settings</h1>
           <div className='NewVice'>
